Moved manager to session/manager
