+++ src/app/modules/pda/exporter/PDAGraphExporter.js
@@ -1,4 +1,4 @@
-import AbstractGraphExporter from 'manager/export/AbstractGraphExporter.js';
+import AbstractGraphExporter from 'session/manager/export/AbstractGraphExporter.js';
 
 import { JSON as JSONGraphParser } from 'modules/pda/graph/PDAGraphParser.js';
 import { downloadText } from 'util/Downloader.js';
