+++ src/app/session/manager/DrawerManager.js
