+++ src/app/session/manager/MenuManager.js
