+++ src/app/session/manager/RenderManager.js
