+++ src/app/session/manager/TooltipManager.js
