+++ src/app/session/manager/ViewportManager.js
