+++ src/app/session/manager/export/AbstractGraphExporter.js
