+++ src/app/session/manager/export/ExportManager.js
