+++ src/app/session/manager/hotkey/HotKeyManager.js
