+++ src/app/session/manager/hotkey/HotKeyView.css
