+++ src/app/session/manager/hotkey/HotKeyView.js
