+++ src/app/session/manager/labeleditor/LabelEditorView.css
