+++ src/app/session/manager/labeleditor/LabelEditorView.js
