+++ src/app/session/manager/undo/AbstractEventHandler.js
