+++ src/app/session/manager/undo/UndoManager.js
